fixes and updates for NetemInterface()
remove references to the sudo_pwd member. it was a bad idea to
begin with. the best way to handle the privileges needed by the
ip and tc commands is to make use of sudoers and sudoers.d

add a DEFAULTS member that provides fall back values for side,
ctrl_fqdn, and ctrl_port. it is somewhat cleaner this way