serve netem control instance(s) for remote access
starts control.NetemInterface() instances and exposes them for remote
access.
for the moment only RPC access using the Pyro4 Python package is enabled.
this is still a very rough first cut with limited access functionality (for
example, a client must run on the localhost in order to access the server)