use pickle as a serializer for the Pyro4 server
we need to pass emulation(s) objects between the client and the server.
Pyro4 will not serialize python classes (or class instances) by default
because of security issues. we replace the default serpent serializer
with pickle. we could have also used the dill serializer but there is
a performance hit associated with it