replace NetemInterface.state attribute with a @property getter
the state property of a NetemInterface object must be available remotely
but Pyro4 does not expose attributes because of security issues.
therefore we will convert all the attributes that need to be visible over
the network with @property decorated methods.
also, it makes sense to return the interface name as part of the state
property.