introduced function _locate_pyro4_ns()
locate a Pyro4 naming server if so specified in the start-up arguments

if there is a Pyro4 naming server running on this subnet, then return a
proxy for it.

note that this function as written and invoked in this module can only
locate name servers that accept the serializer specified in the config
module. by default the pyro4 name server (pyro4-ns) doesn't accept
the insecure serializers required by this module.